apps/web/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
alembic
databases
httpx
pandas
numpy
python-dotenv
openai
google-generativeai
